# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Football Training Board API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  clubName: string;
  /** @nullable */
  clubCity?: string | null;
  /** @nullable */
  clubCountry?: string | null;
}

export interface LoginRequest {
  email: string;
  password: string;
  section?: string;
}

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  createdAt: string;
}

export interface Club {
  id: number;
  name: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  foundedYear?: number | null;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  club: Club;
  role: string;
}

export interface UpdateClubRequest {
  name?: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  foundedYear?: number | null;
  /** @nullable */
  description?: string | null;
}

export interface ClubMember {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  role: string;
  joinedAt: string;
  clubSection?: string[] | null;
  staffRole?: string | null;
  registered?: boolean | null;
  registrationNumber?: string | null;
  phone?: string | null;
  licenseType?: string | null;
  specialization?: string | null;
  degreeScienzeMoto?: boolean | null;
  degreeScienzeMotoType?: string | null;
  teamAssignments?: { id: number; teamId: number; teamName: string }[];
}

export interface InviteMemberRequest {
  email: string;
  firstName: string;
  lastName: string;
  password: string;
  role: string;
  clubSection?: ("scuola_calcio" | "settore_giovanile" | "prima_squadra")[];
  registered?: boolean;
  registrationNumber?: string;
  phone?: string;
  licenseType?: string;
  specialization?: string;
  degreeScienzeMoto?: boolean;
  degreeScienzeMotoType?: string;
}

export interface UpdateMemberRoleRequest {
  firstName?: string;
  lastName?: string;
  email?: string;
  newPassword?: string;
  role: string;
  clubSection?: ("scuola_calcio" | "settore_giovanile" | "prima_squadra")[];
  staffRole?: string;
  registered?: boolean;
  registrationNumber?: string;
  phone?: string;
  licenseType?: string;
  specialization?: string;
  degreeScienzeMoto?: boolean;
  degreeScienzeMotoType?: string;
  teamIds?: number[];
}

export interface TrainingSlot {
  day: string;
  startTime: string;
  endTime: string;
}

export interface Team {
  id: number;
  clubId: number;
  name: string;
  /** @nullable */
  ageGroup?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  coachId?: number | null;
  /** @nullable */
  coachName?: string | null;
  playerCount: number;
  createdAt: string;
  /** @nullable */
  trainingSchedule?: TrainingSlot[] | null;
  /** @nullable */
  clubSection?: string | null;
}

export interface CreateTeamRequest {
  name: string;
  /** @nullable */
  ageGroup?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  coachId?: number | null;
  /** @nullable */
  trainingSchedule?: TrainingSlot[] | null;
  clubSection?: "scuola_calcio" | "settore_giovanile" | "prima_squadra";
}

export interface UpdateTeamRequest {
  name?: string;
  /** @nullable */
  ageGroup?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  coachId?: number | null;
  /** @nullable */
  trainingSchedule?: TrainingSlot[] | null;
}

export interface Player {
  id: number;
  clubId: number;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  firstName: string;
  lastName: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  nationality?: string | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  jerseyNumber?: number | null;
  status: string;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreatePlayerRequest {
  firstName: string;
  lastName: string;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  nationality?: string | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  jerseyNumber?: number | null;
  status: string;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdatePlayerRequest {
  firstName?: string;
  lastName?: string;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  nationality?: string | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  jerseyNumber?: number | null;
  status?: string;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface TrainingSession {
  id: number;
  clubId: number;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  scheduledAt: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  location?: string | null;
  status: string;
  /** @nullable */
  objectives?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreateTrainingSessionRequest {
  /** @nullable */
  teamId?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  scheduledAt: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  location?: string | null;
  status: string;
  /** @nullable */
  objectives?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateTrainingSessionRequest {
  /** @nullable */
  teamId?: number | null;
  title?: string;
  /** @nullable */
  description?: string | null;
  scheduledAt?: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  location?: string | null;
  status?: string;
  /** @nullable */
  objectives?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface DashboardStats {
  totalTeams: number;
  totalPlayers: number;
  totalMembers: number;
  upcomingTrainingSessions: number;
  recentTrainingSessions: TrainingSession[];
}

export type ListPlayersParams = {
  teamId?: number;
};

export type ListTrainingSessionsParams = {
  teamId?: number;
};